requirements/base.txt
# django-rosetta
# django-sendfile
# django-widget-tweaks
# django-stdimage
psycopg2
django-compressor
django-cors-headers==2.0.2
django-model-utils
django-registration-redux
djangorestframework
django-silk
Django>=1.10,<1.11
ipython
python-memcached
pytz
raven
Sphinx
sphinx-rtd-theme
pytest-django
django-environ
uwsgi